remove old navigation bar and rename cupertino and create command
