+++ lib/Scenes/Page/ProfilePage/profile_page.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+// Dummy Page
 class ProfilePage extends StatefulWidget {
   const ProfilePage({super.key});
