+++ lib/Scenes/Page/main_page.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:timeboxing/Scenes/Page/HomePage/home_page.dart';
-import 'package:timeboxing/Shared/Widget/cupertino.dart';
+import 'package:timeboxing/Shared/Widget/bottom_navigation_bar.dart';
+import 'package:timeboxing/Shared/Widget/bottom_navigation_bar.dart';
 
 class MainPage extends StatelessWidget {
   const MainPage({super.key});
@@ -9,7 +10,7 @@ class MainPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return const SafeArea(
       child: Scaffold(
-        bottomNavigationBar: CupertinoTabBarApp(),
+        bottomNavigationBar: TimeboxingBottomNavigationBar(),
         body: HomePage(),
       ),
     );
